requirements.txt
# Core dependencies
streamlit==1.39.0
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.25.0

# AI Models
openai>=1.40.0
google-generativeai>=0.8.0

# Data & APIs
yfinance==0.2.44
alpha-vantage==2.3.1
newsapi-python==0.2.7
requests==2.32.3

# Data analysis
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1

# Visualization
plotly==5.24.1
matplotlib==3.9.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2
